Move files and add common func to validate auth
