+++ fed_mng/socketio/site_admin.py
@@ -1,35 +1,18 @@
-from typing import Literal
+from typing import Any, Literal
 
-from flaat.exceptions import FlaatUnauthenticated
 from socketio import AsyncNamespace
 
-from fed_mng.auth import has_role
+from fed_mng.socketio.utils import validate_auth_on_connect
 
 
 class SiteAdminNamespace(AsyncNamespace):
-    async def on_connect(self, sid, environ, auth: dict[Literal["token"], str]):
-        """When connecting evaluate user authentication.
-
-        Invoke OPA to check if user is authenticated or not and check it has the
-        site_admin role.
-
-        Args:
-            sid (_type_): _description_
-            environ (_type_): _description_
-        """
-        print("Connecting to namespace:", self.namespace, sid, environ, auth)
-        if auth is None or auth.get("token", None) is None:
-            raise ConnectionRefusedError(
-                "Authentication failed: No authentication data nor access token received."
-            )
-        try:
-            assert has_role(
-                auth.get("token", ""), self.namespace[1:]
-            ), ConnectionRefusedError(
-                "Authentication failed: User does not have needed access rights."
-            )
-        except FlaatUnauthenticated as e:
-            raise ConnectionRefusedError(e) from e
+    async def on_connect(self, sid: str, environ: dict[str, Any], auth: dict[Literal["token"], str]):
+        """When connecting evaluate user authentication."""
+        print(f"Connecting to namespace: {self.namespace}")
+        print(f"SID: {sid}")
+        print(f"Environment variables: {environ}")
+        print(f"Auth data: {auth}")
+        validate_auth_on_connect(auth=auth, target_role=self.namespace[1:])
         print(f"Connected to namespace '{self.namespace}' with sid '{sid}'")
 
     async def on_disconnect(self, sid):
